SLIM-1668 refactor Scan M-Bus channels and Couple M-Bus functionality
